cast to memory size for correct comparison on 64bit

+++ Hash.h
@@ -81,7 +81,7 @@ namespace Hash {
     **/
     template<std::size_t N = 20, typename T>
         requires(std::is_integral_v<T> &&
-                 (N > 0 && N < std::numeric_limits<T>::digits - 1))
+                 (N > 0 && N < static_cast<std::size_t>(std::numeric_limits<T>::digits) - 1))
     constexpr T hash_coordinate_to_integral(T x, T y, T z) {
         constexpr T A{ 73856093 };
         constexpr T B{ 19349663 };
@@ -90,7 +90,7 @@ namespace Hash {
     }
     template<std::size_t N = 20, typename T>
         requires(std::is_integral_v<T> &&
-                 (N > 0 && N < std::numeric_limits<T>::digits - 1))
+                 (N > 0 && N < static_cast<std::size_t>(std::numeric_limits<T>::digits) - 1))
     constexpr auto hash_coordinate_to_integral(T x, T y) {
         return hash_coordinate_to_integral<N>(x, y, x ^ y);
     }
@@ -113,13 +113,13 @@ namespace Hash {
         constexpr T A{ 73856093 };
         constexpr T B{ 19349663 };
         constexpr T C{ 83492791 };
-        assert(N > 0 && N < std::numeric_limits<T>::digits - 1);
+        assert(N > 0 && N < static_cast<std::size_t>(std::numeric_limits<T>::digits) - 1);
         return ((1 << N) - 1) & (x * A ^ y * B ^ z * C);
     }
     template<typename T>
         requires(std::is_integral_v<T>)
     constexpr T hash_coordinate_to_integral(std::size_t N, T x, T y) {
-        assert(N > 0 && N < std::numeric_limits<T>::digits - 1);
+        assert(N > 0 && N < static_cast<std::size_t>(std::numeric_limits<T>::digits) - 1);
         return hash_coordinate_to_integral(N, x, y, x ^ y);
     }
